- Document Location, Expression, QueryResult (#109)
